requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1

# Vector database
chromadb==0.4.18

# Graph database
neo4j>=5.0.0

# OpenAI integration
openai>=1.10.0

# Vector operations and embeddings
numpy>=1.21.0
sentence-transformers>=2.0.0
langchain>=0.1.0
langchain-openai>=0.0.8

# Text processing and chunking
# tiktoken==0.5.2  # Temporarily disabled due to Rust compilation

# Environment and configuration
python-dotenv==1.0.0
pydantic>=2.5.2
pydantic-settings>=2.3.0

# Utilities
httpx==0.25.2
python-dateutil==2.8.2

# Real-time and async support for Phase 3
asyncio-mqtt==0.16.1
redis==5.0.1
celery==5.3.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Logging
structlog==23.2.0 